requirements.txt
streamlit==1.25.0
gradio==3.39.0
huggingface-hub==0.16.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
textblob==0.17.1
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
